feature(all):完善阶段6的开发

+++ internal/model/refresh_token.go
@@ -7,7 +7,7 @@ import (
 // RefreshToken Refresh Token 模型
 type RefreshToken struct {
 	ID        uint      `gorm:"primaryKey" json:"id"`
-	UserID    uint      `gorm:"not null;index" json:"user_id"`
+	UserID    uint64    `gorm:"not null;index" json:"user_id"`
 	Token     string    `gorm:"type:varchar(64);not null;uniqueIndex" json:"token"`
 	ExpiresAt time.Time `gorm:"not null;index" json:"expires_at"`
 	CreatedAt time.Time `gorm:"autoCreateTime" json:"created_at"`
